feat: make a thread safe version of general store
Note: I had to turn off the HasStore for General Store due to
conflicting implementations, I'd like to work out how to bring that back
