+++ storage/src/composite_store/store/mod.rs
@@ -1,0 +1,5 @@
+mod general_store;
+pub use general_store::*;
+
+mod thread_safe_general_store;
+pub use thread_safe_general_store::*;
